+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This file is following [keep-a-changelog](https://keepachangelog.com).
 ## [Unreleased](https://github.com/kupl/apps-sal)
 
 ### Added
+- List utility functions (`map` and `filter`) for processing `Dataset`.
 - Implement exporting dataset as `jsonl` format.
 
 ### Changed